+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ehollidg <ehollidg@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/03/20 12:52:05 by ehollidg       #+#    #+#                */
-/*   Updated: 2019/03/27 17:17:55 by ehollidg      ########   odam.nl         */
+/*   Updated: 2019/04/08 15:47:59 by ehollidg      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t		ft_strlen(char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
